Updates readme. Commits .env.
The rationale about committing .env:
This setup works out of the box in any environment and shoud do so without any effort required from the user.
The .env file isn't acutally used to be able to deploy into different environments, but to have have all configuration options for docker-compose in one place.
grafana_port could be defaulted in docker-compose.yml. Unfortunately I found no way to do this with COMPOSE_PROJECT_NAME. So committing the .env file is the simplest option that still allows to add easily add additional .env files if needed.